refactor(audio): tidy tone() module layout
- Move `getAudioContext` up next to `init` so the procedural-audio
  surface introduces itself at the top of the file rather than
  trailing the file-based playback section.
- Extract `ToneOptions` into `audio/types.ts` — keeps `audio.ts`
  focused on runtime; re-exported from `audio.ts` so
  `me.audio.ToneOptions` resolves as before.
- Strip implementation-detail mentions of Howler from JSDoc — users
  see "the audio module" / "the shared AudioContext", not the wrapped
  library name.
- Clarify on `tone`: WebAudio is required; without it the call is a
  silent no-op (`getAudioContext()` returns `null`).

No behavioural change.
